codeismine/requirements.txt
# Core
Django==5.1.3
django-configurations==2.5.1
gunicorn==23.0.0
newrelic==10.3.0
setuptools==75.6.0

# For the persistence stores
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Model Tools
django-model-utils==5.0.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.15.2
django-filter==24.3

# Developer Tools
ipdb==0.13.13
ipython==8.29.0
mkdocs==1.6.1
flake8==7.1.1

# Testing
mock==5.1.0
factory-boy==3.3.1
pytest-django==4.9.0
coverage==7.6.7

# Static and Media Storage
django-storages==1.14.4
boto3==1.35.66

django-cors-headers==4.6.0
